+++ lib/pry/repl.rb
@@ -179,7 +179,7 @@ class Pry
           end
         end
 
-        if input == Readline
+        if defined?(Readline) and input == Readline
           input_readline(current_prompt, false) # false since we'll add it manually
         elsif defined? Coolline and input.is_a? Coolline
           input_readline(current_prompt)